Freezing Docker installation script
Since the beginning of July 2017, the installation script served via get.docker.com invokes systemctl to start the docker daemon.
We do not need the docker daemon to run in the container (as we use the one on the host), but the build of the image fails as there is not systemctl in the container (and it is a nightmare to have).
We resort to the old script (without systemctl) until it stops working.