Migrate JupyterHub image to Alma9
This creates a slimmer version of the former JupyterHub image
that inherits from an Alma9 base image.

The main benefits of this change are:
- We execute on a newer OS with a newer Python version (3.9)
- We reduce the number of packages that are installed to the
minimum necessary.

The versions of the packages we install remain the same, notably
JupyterHub (1.4.2).

This is just a first step towards the update of JupyterHub and its
related packages.