Install helm v2 temporarily for compatibility with current k8s Spark cluster
The current k8s Spark cluster is v1.18.6 and it is not compatible
with helm v3 -- some commands in sparkk8s_token.sh fail.
For now, install a helm v2 binary instead until the k8s Spark
cluster is upgraded. When that happens, revert this commit and
update sparkk8s_token.sh to make it compatible with helm v3.