Remove delete_ticket.sh script, not needed anymore
The deletion of the EOS ticket secret is already done during
the stop method of our k8s spawner (this is included in the JH4
upgrade changes). Since the culler asks the hub to cull sessions,
the stop method will be called for each of those sessions, and
thus the EOS ticket secret will be removed.